Added BellmanFord and PrimsAlgo
BellmanFord is used to find the shortest path from a source vertex to all other vertices in a weighted graph. It is similar to Dijkstra's algorithm but it can work with graphs in which edges can have negative weights.
PrimsAlgo is used to find the minimum spanning tree of a weighted undirected graph. It finds a subset of the edges that forms a tree that includes every vertex, where the total weight of all the edges in the tree is minimized.